requirements.txt
asgiref==3.2.10
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.1
djangorestframework==3.11.1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hibagent==1.0.1
hpack==3.0.0
hstspreload==2020.9.23
httpcore==0.9.1
httplib2==0.14.0
httpx==0.13.3
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
joblib==0.16.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.22.0
requests-unixsocket==0.2.0
rfc3986==1.4.0
scikit-learn==0.23.2
scipy==1.5.2
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sniffio==1.1.0
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
threadpoolctl==2.1.0
tqdm==4.49.0
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1
